refactor(wordpress): extract Playground bootstrap helpers into shared lib
Pulls the diagnostics + four shared bootstrap stages out of
`scripts/test/playground-runner.php` into a new shared library at
`scripts/lib/playground-bootstrap.php` so future runner kinds
(benchmarks, integration tests, ad-hoc Playground dispatchers) can
boot WordPress through the exact same code as the test runner.

Why now
-------

A drift between "the boot path tests run against" and "the boot path
bench runs against" silently invalidates every cross-runner
comparison. A regression in `boot` or `install` would surface as a
test slowdown but be invisible to bench (or vice versa) — and we'd
chase the wrong cause for hours. Single-source-of-truth for
bootstrap is the only honest path.

This refactor lands ahead of the bench dispatcher PR (Phase 4 of
#214) so the bench runner can `require_once` this lib and inherit
the test runner's boot path verbatim, with no copy-paste.

What moved (no behavior change)
-------------------------------

Helpers (line 52-94 of the original):
  - `pg_log()` — append a line to the structured log
  - `pg_stage_begin()` / `pg_stage_ok()` / `pg_stage_fail()` — stage
    lifecycle markers
  - Warning/notice handler + shutdown handler now wrapped in
    `pg_install_diagnostics_handlers()` so callers can opt in
    explicitly. Previously installed via top-level side-effect; now
    installed via single function call. Net behaviour identical.

Stage executors (lines 107-265 of the original):
  - `pg_run_boot_stage()` — render `wp-tests-config.php`, load
    composer autoload. Returns the config path so callers don't
    have to hard-code `/tmp/wp-tests-config.php`.
  - `pg_run_install_stage(array $cfg)` — invoke wp-phpunit's
    `install.php` in-process. Takes `config_path` from the boot
    stage (and an optional `tests_dir` override).
  - `pg_run_load_deps_stage(array $cfg)` — load every Plugin-Name-
    headed entry file in the declared dependency mounts.
  - `pg_run_load_component_stage(array $cfg)` — load the plugin or
    theme under test. The `db.php` skip rule and theme-detection
    via `style.css` are unchanged.

What stayed in playground-runner.php
------------------------------------

PHPUnit-specific stages remain inline:

  - `load_fixtures` — wp-phpunit testcase classes, mock mailer,
    harness filters. Bench has no use for these.
  - `discover_tests`, `pg_parse_phpunit_config`, `pg_discover_tests` —
    PHPUnit XML parsing + test file walking. Bench has its own
    workload-discovery shape (`tests/bench/*.php`).
  - `load_tests`, `run_tests` — `require` test files + `PHPUnit\TextUI\TestRunner`
    invocation. Bench replaces these with an iteration loop +
    percentile aggregation.

Global contract
---------------

The library reads the global `$result_file` and `$current_stage`
that callers set up before requiring it. Keeping globals (rather
than threading the result-file path through every helper signature)
preserves byte-identical call sites in `playground-runner.php` —
the diff is pure relocation, not an API redesign.

Verification
------------

`scripts/test/playground-dropin-smoke.sh` passes against the
refactored runner:

```
Running PHPUnit tests via WordPress Playground...
  Plugin: dropin-coexistence
  Backend: playground (PHP-WASM + SQLite)
PHPUnit 9.6.34 by Sebastian Bergmann and contributors.

...                                                                 3 / 3 (100%)
Time: 00:00.069, Memory: 48.50 MB
OK (3 tests, 8 assertions)

✓ Drop-in coexistence smoke test PASSED
```

Same `STAGE_BEGIN` / `STAGE_OK` lines, same `PLUGIN_DETECTED` /
`NOTICE:` lines, same `OK (n tests, m assertions)` from PHPUnit.

Plus a structural smoke confirms all nine extracted symbols
(`pg_log`, `pg_stage_begin`, `pg_stage_ok`, `pg_stage_fail`,
`pg_install_diagnostics_handlers`, `pg_run_{boot,install,load_deps,load_component}_stage`)
are reachable from the lib and emit the same log lines they did
when defined inline.

Diff stat: 1 file changed, +19/-162 lines moved into a new
257-line lib with focused docblocks per symbol.

AI assistance
-------------

- **AI assistance:** Yes
- **Tool(s):** Claude Code (Opus 4.7)
- **Used for:** Read every stage and helper in
  `scripts/test/playground-runner.php` to map the extraction
  boundary; wrote the lib; refactored the runner to consume it;
  ran the dropin-coexistence smoke + a structural pre-flight
  smoke to verify behaviour parity. Chris pushed back on an
  initial plan to duplicate the boot stages into a parallel
  bench template, which forced the cleaner refactor-first
  approach this PR represents.
